Redid the project lol, rough framework done
