--- README.md
@@ -1,2 +1,0 @@
-# LeagueOfLegendsBot
-Bot for league match histories and more for friends
