+++ lolbot/static/css/main.css
@@ -1,0 +1,6 @@
+body {
+    background-color: #f0f0f0;
+    font-family: 'Roboto', sans-serif;
+    margin: 0;
+    padding: 0;
+}
