+++ lolbot/templates/index.html
@@ -1,0 +1,9 @@
+{% extends 'base.html' %}
+
+{% block head %}
+<h1>League of Legends Bot</h1>
+{% endblock %}
+
+{% block body %}
+<h2>Enter your summoner name:</h2>
+{% endblock %}
