+++ pyproject.toml
@@ -1,0 +1,19 @@
+[tool.poetry]
+name = "lolbot"
+version = "0.1.0"
+description = ""
+authors = ["kvta1 <kevintran@live.nl>"]
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.10"
+Flask = "^3.0.3"
+numpy = "^2.1.0"
+pandas = "^2.2.2"
+requests = "^2.32.3"
+python-dotenv = "^1.0.1"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
