--- src/testing/matches_df.csv
@@ -1,3 +1,0 @@
-0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19
-EUW1_6673018592,EUW1_6672509981,EUW1_6672472548,EUW1_6672438811,EUW1_6672401362,EUW1_6672359878,EUW1_6672316852,EUW1_6672060274,EUW1_6671795900,EUW1_6671767496,EUW1_6671704719,EUW1_6671625333,EUW1_6671612651,EUW1_6670970453,EUW1_6670939636,EUW1_6670357016,EUW1_6670344696,EUW1_6670308282,EUW1_6670250184,EUW1_6669113923
-EUW1_6672932300,EUW1_6672509981,EUW1_6672472548,EUW1_6672438811,EUW1_6672401362,EUW1_6672359878,EUW1_6672316852,EUW1_6672280058,EUW1_6672122023,EUW1_6672060274,EUW1_6671795900,EUW1_6671767496,EUW1_6671716788,EUW1_6670425441,EUW1_6670034057,EUW1_6670022378,EUW1_6670020592,EUW1_6670001820,EUW1_6669830821,EUW1_6669787329
