--- src/testing/summoners_data.csv
@@ -1,3 +1,0 @@
-name,id,accountId,puuid,profileIconId,revisionDate,summonerLevel
-Tampai,kb6I3_xf83HYAMwRcKdl3BkvmxRKiyfbV9XJ0UnbD3tJEws,6JLFhciTczGvDdSILUy29dyfCr_wNZxFWmY4x06L1jp5qA,Lf0ghEcdrzd-zmXH3pt3bzhBejqrnGXFPgLRLpA0dKAD5mK1YyTld9kHqJ59mA_8bNjPk0C9QSEmeg,1407,1699790029757,346
-VinhQ,9HHQ1XHbxcdMDGp7k3WaoVVSI1RRJK-HdXHQSjiMBGQjYXA,F0RJyP8rXH6rYyLTSCi18gzMW2hHWvEqAKEQjLXOpZjvvA,HgxxEH6di_UMvSQ76RtaTIBFXelYZROls2H0AdWstUNcUpwPfqU_W0v74_Axm6SP1AL15E1uYULQeg,593,1699785723623,667
