+++ templates/index.html
@@ -1,0 +1,21 @@
+{% extends 'base.html' %}
+
+{% block head %}
+<title>The ULTIMATE League of Legends Stats Resource 🥵🥵</title>
+{% endblock %}
+
+{% block body %}
+<div class="content">
+    <h1 style="text-align: center">The ULTIMATE League of Legends Stats Resource 🥵🥵</h1>
+    <h2 style="text-align: center">Enter your summoner name</h2>
+
+
+
+    <div class="form">
+        <form action="/" method="POST">
+            <input type="text" name="content" id="content">
+            <input type="submit" value="Name">
+        </form>
+    </div>
+</div>
+{% endblock %}
